fix: 修复审计发现的4个严重+6个高危漏洞
严重：
- llm_service.chat 统一注入配额检查+调用记录（无需逐路由添加）
- 中间件实时检查 approval_status+billing_due_date（活跃会话不再绕过）
- refresh 端点增加审批/计费/租户状态验证

高危：
- billing_service 使用 SubscriptionStatus Enum 替代原始字符串
- billing_amount 统一为 Float（消除 Integer/Float 类型不匹配）
- 中间件 NULL billing_due_date 防御性检查（仅已配置计费的租户检查）
- 测试 mock 补充 approval_status/is_platform_admin 字段

+++ app/models/tenant.py
@@ -1,7 +1,7 @@
 """
 租户（Tenant）模型 - SaaS 多租户核心
 """
-from sqlalchemy import Column, String, DateTime, Boolean, Text, Integer, Enum as SQLEnum
+from sqlalchemy import Column, String, DateTime, Boolean, Text, Integer, Float, Enum as SQLEnum
 from sqlalchemy.orm import relationship
 from datetime import datetime
 import enum
@@ -70,7 +70,7 @@ class Tenant(Base):
     approved_by = Column(String(200), nullable=True)
     rejection_reason = Column(Text, nullable=True)
     billing_cycle = Column(String(20), default="monthly")  # monthly/yearly
-    billing_amount = Column(Integer, default=0)  # 月/年费（元）
+    billing_amount = Column(Float, default=0)  # 月/年费（元）
     billing_due_date = Column(DateTime, nullable=True)
     billing_method = Column(String(50), nullable=True)  # wechat/alipay/bank_transfer
     registered_from = Column(String(500), nullable=True)  # 注册来源（IP/域名）
